% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfgs.R
\name{bfgs}
\alias{bfgs}
\title{Broyden Fletcher Goldfarb Shanno method.}
\usage{
bfgs(r, v, hess)
}
\arguments{
\item{r}{A vector, the difference of the gradient values at the points (dfx) of the current and previous iteration.}

\item{v}{A vector, the difference of the points (x) of the current and previous iteration.}

\item{hess}{A matrix, hessian.}
}
\value{
C, a approximation of the inverse hessian.
}
\description{
\code{bfgs} is a iterative method for update the approximate hessian (or its inverse) in Quasi-Newton method.
}
\examples{
r <- df_k - df_k1
v <- x_k - x_k1
hess <- diag(length(x_k))
hess <- hess + bfgs(r, v, hess)

}
\references{
\enumerate{
\item Wikipedia, \emph{Quasi-newton method} \url{https://en.wikipedia.org/wiki/Quasi-Newton_method}.
\item Ramirez, Jaime A.; Campelo, Felipe; Guimaraes, Frederico G.; Batista, Lucas S.; Takahashi, Ricardo H. C.;
\emph{Notas de aula de Otimizacao}, pages 26:28.
}
}
\seealso{
The documentation of the Quasi-newton method \code{quasiNewton} in this package.
}
