% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parameters.R
\name{check_parameters}
\alias{check_parameters}
\title{Check input parameters on quasi-Newton}
\usage{
check_parameters(obj.list, x.list, ...)
}
\arguments{
\item{obj.list}{Either a function or a list with the following names \cr
f: the objective function \cr
df: its gradient (optional; where the numerical version is used by default)
\cr}

\item{x.list}{Either a vector with initial solution or a list with the
following names \cr
  x: a vector with its value in the search space \cr
  fx: a scalar with its objective value \cr
  dfx: a vector with its gradient value \cr}

\item{...}{Optional parameters passed to the \code{gradient} function}
}
\value{
Returns two lists: \cr
  \code{obj.list}: a list with names \code{f} and \code{df} with the
  objective function and its gradient, respectively \cr
  \code{x.list}: a list with the names \code{x}, \code{fx} and \code{dfx}
  with the initial point properly evaluated.
}
\description{
The quasiNewton algorithm can be run by providing different types of input,
but they all need to be accomodated to a common format. This is performed
in this function
}
