% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkparameters.R
\name{checkparameters}
\alias{checkparameters}
\title{Check parameter (Options.list)}
\usage{
checkparameters(obj.list, x.list, Options.list)
}
\arguments{
\item{Options.list}{A list}
}
\value{
Returns a complete Options.list with default values.
}
\description{
\code{checkparameters} is a function that checks if the parameter Options.list is complete
to use in some functions in this package.
}
\section{About the parameter Options.list}{

This list contains all parameters that is not fundamental for the function. In other words,
Options.list receives instances that have default values in all functions that are used,
but the user can have a greater control of the operation of the code, if desired.
\itemize{
\item use_bracketing = TRUE or FALSE, used in function \code{linesearch}
\item step_bracketing, (number), used in function \code{linesearch}
\item A and B (numbers), used in function \code{linesearch}
\item Method ("quadraticinterpolation" or "goldensection"), used in function \code{linesearch}
\item rhoBacktracking e cBacktracking, used in function \code{linesearch}
\item Eps, epsilon, error tolerance, default = 1e-6
\item Eps_f for the functions \code{stopping_conditions}, \code{quasiNewton} and \code{conjugateGradient}
\item Eps_df for the functions \code{stopping_conditions}, \code{quasiNewton} and \code{conjugateGradient}
\item maxNI maximum number iterations for the functions \code{stopping_conditions}, \code{quasiNewton} and
\code{conjugateGradient}
}
}

