% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Shortcut to evaluate a new point at an objective function and its gradient}
\usage{
evaluate(obj.list, x)
}
\arguments{
\item{obj.list}{A list with the problem functions}

\item{x}{A vector solution to be evaluated}
}
\value{
A list with x evaluated at \code{f} and \code{df}
}
\description{
Shortcut to evaluate a new point at an objective function and its gradient
}
