% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldensection.R
\name{goldensection}
\alias{goldensection}
\title{Golden Section Method.}
\usage{
goldensection(phif, a = 0, b = 1, eps = 1e-06)
}
\arguments{
\item{phif}{A univariate function.}

\item{a, b}{Numbers, initial interval.}

\item{eps}{A number, a tolerance to stop the optimization when the current
interval becomes smaller.}
}
\value{
Returns the step length \code{alpha}.
}
\description{
\code{goldensection} is a line search method that reduces the initial interval
to get a optimum with the highest precision needed. In this implementation the reduction
of the interval, when it happens, occurs by dividing by half.
}
\section{About the parameter phif}{

\code{phif} should be a univariate function that returns the value of the function
at a given point.
}

\examples{
phi <- function(alpha) {...}
goldensection(phi, a = 0, b = 10, eps = 1e-4)
goldensection(phi, 3, 7, 1e-5)
}
\references{
\enumerate{
\item Rao, Singiresu S.; \emph{Engineering Optimization Theory and and Practice}, 4th ed., page 267.
\item Ramirez, Jaime A.; Campelo, Felipe; Guimaraes, Frederico G.; Batista, Lucas S.; Takahashi, Ricardo H. C.;
\emph{Notas de aula de Otimizacao}, pages 10:13.
}
}
\seealso{
The documentation of the function \code{univariate_f} in this package.
}
