% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Gradient by the numerical approximation}
\usage{
gradient(f, x, fx = NULL, dx = 1e-06, method = c("ffd", "cfd"))
}
\arguments{
\item{f}{A function, representing the objective function.}

\item{x}{A number or vector with length n, indicating the current point.}

\item{fx}{A number, the objective value calculated at x.}

\item{dx}{A number, the small perturbation in x.}

\item{method}{character string, specifying the discretization method. \cr
"ffd": Forward finite differences. Requires n+1 evaluations (or n if fx is
provided) \cr
"cfd": Central (or Symmetrical) finite differences. Requires 2n evaluations
(fx is not used)}
}
\description{
\code{gradient} is a function based on the algorithm of numerical
differentiation for estimating the gradient using finite difference
approximations.
}
\references{
\enumerate{
\item Wikipedia, \emph{Numerical differentiation}, \url{https://en.wikipedia.org/wiki/Numerical_differentiation}.
}
}
