% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normE.R
\name{normE}
\alias{normE}
\title{Euclidean norm.}
\usage{
normE(vec)
}
\arguments{
\item{vec}{A vector.}
}
\value{
Returns the Euclidean norm of \code{vec}.
}
\description{
Euclidean norm.
}
\examples{
normE(c(-1,2))
normE(c(10, 25))
}
