% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadraticinterpolation.R
\name{quadraticinterpolation}
\alias{quadraticinterpolation}
\title{Quadratic Interpolation Method.}
\usage{
quadraticinterpolation(phi.f, l = 0, u = 1, eps = 1e-06)
}
\arguments{
\item{l, u}{Numbers, initial interval (or points).}

\item{eps}{A number, used in the stop condition.}

\item{phif}{A univariate function.}
}
\value{
Returns the step length \code{alpha}.
}
\description{
\code{quadraticinterpolation} is a one-dimensional minimization method that finds the minimizing
step length \code{alpha} through the approximation by a quadratic function. If the \code{alpha}
is not sot sufficiently to the minimum a refit of the function is used.
}
\section{Reduction of points, interval}{

To reduce the interval between the three points, first evaluate the value of the function
in alpha and in the poinst: lower (\code{l}), upper (\code{u}) and middle (\code{m}).
After this evaluation, it is analyzed (through comparisons) at which two points the
alpha is between and are chosen the next values for the refite
}

\section{About the parameter phif}{

\code{phif} should be a univariate function that returns the value of the function
at a given point.
}

\examples{
phi <- function(alpha) {...}
quadraticinterpolation(phi, a = 0, b = 10, eps = 1e-4)
quadraticinterpolation(phi, 3, 7, 1e-5)
}
\references{
\enumerate{
\item Rao, Singiresu S.; \emph{Engineering Optimization Theory and and Practice}, 4th ed., pages 273:279.

}
}
\seealso{
The documentation of the function \code{univariate_f} in this package.
}
