% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasiNewton.R
\name{quasiNewton}
\alias{quasiNewton}
\title{Unconstrained optimization with Quasi-Newton}
\usage{
quasiNewton(obj.list, x.list, maxNI = 50, eps.df = 1e-06, eps.x = 1e-06,
  alpha0 = 1, rho = 0.5, c = 1e-04, ...)
}
\arguments{
\item{obj.list}{Either an objective function or a list with the following
names \cr
  f: objective function
  df: gradient of f (it not provided, defaults to the numerical version)}

\item{x.list}{Either a vector with an initial solution or a list with the
following names \cr
  x: a vector with its value in the search space \cr
  fx: a scalar with its objective value \cr
  dfx: a vector with its gradient value \cr}

\item{maxNI}{maximum number of iterations}

\item{eps.df}{tolerance in the norm of the gradient}

\item{eps.x}{tolerance in the norm of the difference between two consecutive
solutions}

\item{alpha0}{Initial step size in the backtracking}

\item{rho}{A constant to reduce alpha in backtracking}

\item{c}{A small constant, control parameter.}
}
\value{
Returns a list with the (approximate) optimum.
}
\description{
\code{quasiNewton} solves a multivariate function using a Quasi-Newton with
BFGS udpate.
}
\section{Todo}{

Maybe include alpha0, rho and c into ... and get the appropriate parameters
in each function?
}

\examples{
# The popular Rosenbrock function
f <- function(x)
{
  x1 <- x[1]
  x2 <- x[2]
  return ( 100*(x2 - x1^2)^2 + (1 - x1)^2 )
}
x0 <- c(-1.2,1) #usual starting point
# Run the Quasi-Newton with default parameters (very close to c(1,1))
x.list <- quasiNewton(f, x0) #x.list$x = c(0.9996996 0.9993989))
# Notice the result can be improved by using "CFD" in the numerical gradient
# instead of "FFD":
x.list <- quasiNewton(f, x0, method = "cfd") #x = c(1,1)

}
\references{
\enumerate{
\item Nocedal, Jorge; Wright, Stephen J.; \emph{Numerical Optimization}, 2nd ed., page 37.
\item Wikipedia, \emph{Quasi-newton method} \url{https://en.wikipedia.org/wiki/Quasi-Newton_method}.
}
}
