% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_f.R
\name{univariate_f}
\alias{univariate_f}
\title{Univariate function}
\usage{
univariate_f(fx, x, searchD)
}
\arguments{
\item{fx}{A objective function.}

\item{x}{A point.}

\item{searchD}{A search direction.}
}
\value{
Returns the step length \code{alpha} for the univariate function.
}
\description{
\code{univariate_f} is a function in therms of \code{alpha} and used in optimization methods.
Usually represented by the greek letter "phi" has the following relation with the objective
function \eqn{\phi(\alpha) = f(x + S*\alpha)}, where S represents the search direction.
}
\examples{
f <- function(x) {... return(x)}
phi <- function(f, (1,2), (-3,2))
phi(alpha)
}
