% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfgs.R
\name{bfgs}
\alias{bfgs}
\title{Broyden Fletcher Goldfarb Shanno update method.}
\usage{
bfgs(x.list, xnew.list, H)
}
\arguments{
\item{x.list}{A list with the current solution}

\item{xnew.list}{A list with the new solution in the same iteration}

\item{H}{Current Hessian inverse approximation}
}
\description{
\code{bfgs} is an update scheme to get a new Hessian inverse approximation in
a Quasi-Newton method.
}
\references{
\enumerate{
\item Wikipedia, \emph{Quasi-newton method} \url{https://en.wikipedia.org/wiki/Quasi-Newton_method}.
\item Ramirez, Jaime A.; Campelo, Felipe; Guimaraes, Frederico G.; Batista, Lucas S.; Takahashi, Ricardo H. C.;
\emph{Notas de aula de Otimizacao}, pages 26:28.
}
}
\seealso{
The documentation of the Quasi-newton method \code{quasiNewton} in
this package.
}
