% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Gradient by the numerical approximation}
\usage{
gradient(x, f, fx = NULL, eps = 1e-04)
}
\arguments{
\item{x}{A point (number or vector).}

\item{f}{A objective function.}

\item{fx}{A number, the value of function in x.}

\item{eps}{A number, for calculating the gradient.}
}
\value{
g the numerical approximation of the gradient.
}
\description{
\code{gradient} is a function based on the algorithm of numerical differentiation for estimating  the gradient
using finite difference approximations.
}
\examples{
f <- functios(x) {... return x}
x1 <- c(2,2)
gradient(x1, f)
gradiente(x1, f, fx = f(x1), eps = 1e-3)

}
\references{
\enumerate{
\item Wikipedia, \emph{Numerical differentiation}, \url{https://en.wikipedia.org/wiki/Numerical_differentiation}.
}
}
