% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Gradient by the numerical approximation}
\usage{
gradient(f, x, fx = NULL, dx = 1e-06, method = c("ffd", "cfd"))
}
\arguments{
\item{f}{A function, representing the objective function.}

\item{x}{A number or vector with length n, indicating the current point.}

\item{fx}{A number, the objective value calculated at x.}

\item{dx}{A number, the small perturbation in x.}

\item{method}{character string, specifying the discretization method. \cr
"ffd": Forward finite differences. Requires n+1 evaluations (or n if fx is
provided) \cr
"cfd": Central (or Symmetrical) finite differences. Requires 2n evaluations
(fx is not used)}
}
\value{
df the numerical approximation of the gradient.
}
\description{
\code{gradient} is a function based on the algorithm of numerical
differentiation for estimating the gradient using finite difference
approximations.
}
\examples{
f <- function(x) {return ( sum(x^2) )}
df_analytical <- function (x) {return (2*x)}
x <- c(1,1) #current point
gradient(f, x) #uses ffd by default
gradient(f, x, method = "cfd")
df_analytical(x)

x <- seq(1,100)
dfx_analytical <- df_analytical(x)
dfx_ffd <- gradient(f, x, method = "ffd")
dfx_cfd <- gradient(f, x, method = "cfd")
norm(dfx_analytical - dfx_ffd, type = "2") #error in the FFD approximation
norm(dfx_analytical - dfx_cfd, type = "2") #error in the CFD approximation

}
\references{
\enumerate{
\item Wikipedia, \emph{Numerical differentiation}, \url{https://en.wikipedia.org/wiki/Numerical_differentiation}.
}
}
